Move snakeize methods to utility module (#1298)
